Add warning for out-of-date NanoDB package
If the last-modified date for any of the sources is higher than the
last-modified date of the JAR produced by maven, then a warning will be
emitted to make sure the user is aware that her/his changes may not be
in effect.

This modification also fails silently, e.g. if `find` is missing, or if
the user doesn't have the right version of `stat` to support the printf
flag.