Add shell-commands to interactive client
NanoDB client now supports shell commands of the form \cmd, which are
handled within the InteractiveClient, not by the server.  There are two
commands:

\help prints out useful help information

\source filename.sql will run all commands contained in the file
filename.sql

There are surely interesting wrinkles to iron out.  For example, if the
sourced file contains an "EXIT" command then it will exit the client.
Also, an error should probably terminate the evaluation of a sourced
file, but currently it does not.