Aggregate function bugfixes
A number of bugfixes caused by some of the changes in the arithmetic
type-coercion rules, and the parser-framework changes.

COUNT(DISTINCT a) wasn't being handled correctly due to an issue in
the NanoSQLTranslator flow-control.

The SumAvg/StdDevVar aggregate implementations are now updated to
produce double-precision results when calculating their values.  The
type-converter had been changed to do integer division when the LHS
and RHS are both integers, and this was breaking tests of these
aggregate functions.