Fix overwritten columns in selclause resultSchema
the `SelectClause.computeSchema` method aggregates a `resultSchema`
member by appending all of the ColumnInfos for each of the selectValues
in the statement.

However, when efforts were made to make Schema less "mutable", I think
a change was overlooked that causes all but the last selectValue's
ColumnInfos to remain in the resultSchema.

I believe this was left undetected because the resultSchema of the
SelectClause isn't really used in many places, except for nested
subqueries. This problem was found by a student, that the first out of
two columns in a nested subquery would "disappear" from the schema.