Documents and code changes for Assignment 3
The design document and information document are included under the doc
directory now.

Added a lot of documentation, particularly to the Schema class.

Made a few changes to the ArithmeticOperator and BooleanOperator
classes' simplify() methods so that we can do some basic
simplifications.

Added the StatisticsUpdater so that it's easier for students to modify
statistics based on selection predicates.

Fixed a bug in SelectClause where ORDER BY clauses were using the
FROM-clause schema rather than the SELECT-clause schema.