Support classes for index scans in queries
The IndexScanNode supports performing index-scans against a table based
on a suitable predicate.

To facilitate use of this plan node, the AnalyzedPredicate and
IndexScanEndpoints classes support analyzing a predicate to determine
whether it is an equality lookup and/or a range scan, and which indexes
on a table may be useful for the scan.

Finally, the IndexInfo class has been updated to report whether an index
is a sequential index or a hash index.